lauryn_hope_tuhaise/requirements.txt
django-bootstrap4
django-tailwind